improved perf. ~10%

+++ src/cc/lib/swcdb/core/tests/sertest.cc
@@ -64,7 +64,7 @@ void test_i16() {
       SWC_ASSERT(len == 0));
   }
   ns = Time::now_ns() - ns;
-  std::cout << "i16 took=" << ns << " avg=" << ns / UINT16_MAX << "\n";
+  std::cout << "i16 took=" << ns << " avg=" << (double)ns / UINT16_MAX << "\n";
 }
 
 void test_i24() {
@@ -89,7 +89,7 @@ void test_i24() {
       SWC_ASSERT(len == 0));
   }
   ns = Time::now_ns() - ns;
-  std::cout << "i24 took=" << ns << " avg=" << ns / UINT24_MAX << "\n";
+  std::cout << "i24 took=" << ns << " avg=" << (double)ns / UINT24_MAX << "\n";
 }
 
 void test_i32() {
@@ -114,7 +114,7 @@ void test_i32() {
       SWC_ASSERT(len == 0));
   }
   ns = Time::now_ns() - ns;
-  std::cout << "i32 took=" << ns << " avg=" << ns / UINT32_MAX << "\n";
+  std::cout << "i32 took=" << ns << " avg=" << (double)ns / UINT32_MAX << "\n";
 }
 
 void test_i64() {
@@ -139,7 +139,7 @@ void test_i64() {
       SWC_ASSERT(len == 0));
   }
   ns = Time::now_ns() - ns;
-  std::cout << "i64 took=" << ns << " avg=" << ns / UINT64_MAX << "\n";
+  std::cout << "i64 took=" << ns << " avg=" << (double)ns / UINT64_MAX << "\n";
 }
 
 void chk_vi24(uint24_t n) {
@@ -172,7 +172,7 @@ void test_vi24() {
     c++;
   }
   ns = Time::now_ns() - ns;
-  std::cout << "vi24 lower took=" << ns << " avg=" << ns / c << " c=" << c << "\n";
+  std::cout << "vi24 lower took=" << ns << " avg=" << (double)ns / c << " c=" << c << "\n";
   c = 0;
   ns = Time::now_ns();
   for(uint24_t n=UINT24_MAX-MAX_CHECKS_VI24+1; n <= UINT24_MAX ;n++) {
@@ -182,7 +182,7 @@ void test_vi24() {
       break;
   }
   ns = Time::now_ns() - ns;
-  std::cout << "vi24 upper took=" << ns << " avg=" << ns / c << " c=" << c << "\n";
+  std::cout << "vi24 upper took=" << ns << " avg=" << (double)ns / c << " c=" << c << "\n";
 }
 
 void chk_vi32(uint32_t n) {
@@ -215,7 +215,7 @@ void test_vi32() {
     c++;
   }
   ns = Time::now_ns() - ns;
-  std::cout << "vi32 lower took=" << ns << " avg=" << ns / c << " c=" << c << "\n";
+  std::cout << "vi32 lower took=" << ns << " avg=" << (double)ns / c << " c=" << c << "\n";
   c = 0;
   ns = Time::now_ns();
   for(uint32_t n=UINT32_MAX-MAX_CHECKS+1; n <= UINT32_MAX ;n++) {
@@ -225,7 +225,7 @@ void test_vi32() {
       break;
   }
   ns = Time::now_ns() - ns;
-  std::cout << "vi32 upper took=" << ns << " avg=" << ns / c << " c=" << c << "\n";
+  std::cout << "vi32 upper took=" << ns << " avg=" << (double)ns / c << " c=" << c << "\n";
 }
 
 void chk_vi64(uint64_t n) {
@@ -257,7 +257,7 @@ void test_vi64() {
     c++;
   }
   ns = Time::now_ns() - ns;
-  std::cout << "vi64 lower took=" << ns << " avg=" << ns / c << " c=" << c << "\n";
+  std::cout << "vi64 lower took=" << ns << " avg=" << (double)ns / c << " c=" << c << "\n";
   c = 0;
   ns = Time::now_ns();
   for(uint64_t n=UINT64_MAX-MAX_CHECKS+1; n<=UINT64_MAX;n++) {
@@ -267,7 +267,7 @@ void test_vi64() {
       break;
   }
   ns = Time::now_ns() - ns;
-  std::cout << "vi64 upper took=" << ns << " avg=" << ns / c << " c=" << c << "\n";
+  std::cout << "vi64 upper took=" << ns << " avg=" << (double)ns / c << " c=" << c << "\n";
 }
 
 void test_str16() {
